+++ src/Exception/RadExceptionInterface.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\RadBundle\Exception;
+
+interface RadExceptionInterface extends \Throwable
+{
+}