+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 *   (feature) Add `CommandHelper`.
+*   (feature) Add generic `TagInterface` + `TagHelper`.
 
 
 7.3.0